+++ Library/Backends/ZendEngine3/Backend.php
@@ -739,20 +739,55 @@ class Backend extends BackendZendEngine2
         }
     }
 
-    public function fetchProperty(Variable $symbolVariable, Variable $variableVariable, $property, $readOnly, CompilationContext $context, $useOptimized = false)
-    {
+    /**
+     * {@inheritDoc}
+     *
+     * @param Variable           $symbolVariable
+     * @param Variable           $variableVariable
+     * @param Variable|string    $property
+     * @param bool               $readOnly
+     * @param CompilationContext $context
+     *
+     * @return void
+     */
+    public function fetchProperty(
+        Variable $symbolVariable,
+        Variable $variableVariable,
+        $property,
+        $readOnly,
+        CompilationContext $context
+    ) {
         $flags = 'PH_NOISY_CC';
         if ($readOnly) {
             $flags .= ' | PH_READONLY';
         }
+
         $variableCode = $this->getVariableCode($variableVariable);
         $symbol = $this->getVariableCode($symbolVariable);
-        //TODO: maybe optimizations (read_nproperty/quick) for thisScope access in NG (as in ZE2 - if necessary)
+
         if ($property instanceof Variable) {
-            $context->codePrinter->output('zephir_read_property_zval('.$symbol.', '.$variableCode.', '.$this->getVariableCode($property).', '.$flags.');');
-        } else {
-            $context->codePrinter->output('zephir_read_property('.$symbol.', '.$variableCode.', SL("'.$property.'"), '.$flags.');');
+            $context->codePrinter->output(
+                sprintf(
+                    'zephir_read_property_zval(%s, %s, %s, %s);',
+                    $symbol,
+                    $variableCode,
+                    $this->getVariableCode($property),
+                    $flags
+                )
+            );
+
+            return;
         }
+
+        $template = 'zephir_read_property(%s, %s, ZEND_STRL("%s"), %s);';
+        /* Are we going to init default object property value? */
+        if ($context->currentMethod && $context->currentMethod->isInitializer()) {
+            $template = 'zephir_read_property_ex(%s, %s, ZEND_STRL("%s"), %s);';
+        }
+
+        $context->codePrinter->output(
+            sprintf($template, $symbol, $variableCode, $property, $flags)
+        );
     }
 
     /**
@@ -852,7 +887,6 @@ class Backend extends BackendZendEngine2
      */
     public function updateProperty(Variable $variable, $property, $value, CompilationContext $context)
     {
-        // TODO(serghei): maybe optimizations as well as above
         $value = $this->resolveValue($value, $context);
 
         if ($property instanceof Variable) {
@@ -868,34 +902,20 @@ class Backend extends BackendZendEngine2
             return;
         }
 
+        $template = 'zephir_update_property_zval(%s, ZEND_STRL("%s"), %s);';
         /* Are we going to init default object property value? */
-        if ($context->currentMethod &&
-            preg_match('/^zephir_init_properties/', $context->currentMethod->getName())) {
-            $context->codePrinter->output(
-                sprintf(
-                    'zend_update_property(Z_OBJCE_P(%s), %s, ZEND_STRL("%s"), %s);',
-                    $this->getVariableCode($variable),
-                    $this->getVariableCode($variable),
-                    $property,
-                    $value
-                )
-            );
-
-            return;
+        if ($context->currentMethod && $context->currentMethod->isInitializer()) {
+            $template = 'zephir_update_property_zval_ex(%s, ZEND_STRL("%s"), %s);';
         }
 
         $context->codePrinter->output(
-            sprintf(
-                'zephir_update_property_zval(%s, ZEND_STRL("%s"), %s);',
-                $this->getVariableCode($variable),
-                $property,
-                $value
-            )
+            sprintf($template, $this->getVariableCode($variable), $property, $value)
         );
     }
 
     public function updateStaticProperty($classEntry, $property, $value, CompilationContext $context)
     {
+        // TODO(serghei): Sort out as well as above
         $value = $this->resolveValue($value, $context);
         $context->codePrinter->output('zephir_update_static_property_ce('.$classEntry.', ZEND_STRL("'.$property.'"), '.$value.');');
     }