+++ ext/stub/properties/extendspublicproperties.zep.c
@@ -55,7 +55,7 @@ zend_object *zephir_init_properties_Stub_Properties_ExtendsPublicProperties(zend
 	{
 		zval local_this_ptr, *this_ptr = &local_this_ptr;
 		ZEPHIR_CREATE_OBJECT(this_ptr, class_type);
-		zephir_read_property(&_0, this_ptr, SL("someArrayWithSubArray"), PH_NOISY_CC | PH_READONLY);
+		zephir_read_property_ex(&_0, this_ptr, ZEND_STRL("someArrayWithSubArray"), PH_NOISY_CC | PH_READONLY);
 		if (Z_TYPE_P(&_0) == IS_NULL) {
 			ZEPHIR_INIT_VAR(&_1$$3);
 			zephir_create_array(&_1$$3, 4, 0);
@@ -75,18 +75,18 @@ zend_object *zephir_init_properties_Stub_Properties_ExtendsPublicProperties(zend
 			zephir_create_array(&_2$$3, 1, 0);
 			add_assoc_stringl_ex(&_2$$3, SL("class"), SL("alert alert-danger"));
 			zephir_array_update_string(&_1$$3, SL("danger"), &_2$$3, PH_COPY | PH_SEPARATE);
-			zend_update_property(Z_OBJCE_P(this_ptr), this_ptr, ZEND_STRL("someArrayWithSubArray"), &_1$$3);
+			zephir_update_property_zval_ex(this_ptr, ZEND_STRL("someArrayWithSubArray"), &_1$$3);
 		}
-		zephir_read_property(&_3, this_ptr, SL("someArrayWithSpecifyKeys"), PH_NOISY_CC | PH_READONLY);
+		zephir_read_property_ex(&_3, this_ptr, ZEND_STRL("someArrayWithSpecifyKeys"), PH_NOISY_CC | PH_READONLY);
 		if (Z_TYPE_P(&_3) == IS_NULL) {
 			ZEPHIR_INIT_VAR(&_4$$4);
 			zephir_create_array(&_4$$4, 3, 0);
 			add_assoc_stringl_ex(&_4$$4, SL("a"), SL("A"));
 			add_assoc_stringl_ex(&_4$$4, SL("b"), SL("B"));
 			add_assoc_stringl_ex(&_4$$4, SL("c"), SL("C"));
-			zend_update_property(Z_OBJCE_P(this_ptr), this_ptr, ZEND_STRL("someArrayWithSpecifyKeys"), &_4$$4);
+			zephir_update_property_zval_ex(this_ptr, ZEND_STRL("someArrayWithSpecifyKeys"), &_4$$4);
 		}
-		zephir_read_property(&_5, this_ptr, SL("someArray"), PH_NOISY_CC | PH_READONLY);
+		zephir_read_property_ex(&_5, this_ptr, ZEND_STRL("someArray"), PH_NOISY_CC | PH_READONLY);
 		if (Z_TYPE_P(&_5) == IS_NULL) {
 			ZEPHIR_INIT_VAR(&_6$$5);
 			zephir_create_array(&_6$$5, 4, 0);
@@ -102,22 +102,22 @@ zend_object *zephir_init_properties_Stub_Properties_ExtendsPublicProperties(zend
 			ZEPHIR_INIT_NVAR(&_7$$5);
 			ZVAL_LONG(&_7$$5, 4);
 			zephir_array_fast_append(&_6$$5, &_7$$5);
-			zend_update_property(Z_OBJCE_P(this_ptr), this_ptr, ZEND_STRL("someArray"), &_6$$5);
+			zephir_update_property_zval_ex(this_ptr, ZEND_STRL("someArray"), &_6$$5);
 		}
-		zephir_read_property(&_8, this_ptr, SL("someEmptyArray"), PH_NOISY_CC | PH_READONLY);
+		zephir_read_property_ex(&_8, this_ptr, ZEND_STRL("someEmptyArray"), PH_NOISY_CC | PH_READONLY);
 		if (Z_TYPE_P(&_8) == IS_NULL) {
 			ZEPHIR_INIT_VAR(&_9$$6);
 			array_init(&_9$$6);
-			zend_update_property(Z_OBJCE_P(this_ptr), this_ptr, ZEND_STRL("someEmptyArray"), &_9$$6);
+			zephir_update_property_zval_ex(this_ptr, ZEND_STRL("someEmptyArray"), &_9$$6);
 		}
-		zephir_read_property(&_10, this_ptr, SL("someGetterSetterArray"), PH_NOISY_CC | PH_READONLY);
+		zephir_read_property_ex(&_10, this_ptr, ZEND_STRL("someGetterSetterArray"), PH_NOISY_CC | PH_READONLY);
 		if (Z_TYPE_P(&_10) == IS_NULL) {
 			ZEPHIR_INIT_VAR(&_11$$7);
 			zephir_create_array(&_11$$7, 1, 0);
 			ZEPHIR_INIT_VAR(&_12$$7);
 			ZVAL_STRING(&_12$$7, "holy");
 			zephir_array_fast_append(&_11$$7, &_12$$7);
-			zend_update_property(Z_OBJCE_P(this_ptr), this_ptr, ZEND_STRL("someGetterSetterArray"), &_11$$7);
+			zephir_update_property_zval_ex(this_ptr, ZEND_STRL("someGetterSetterArray"), &_11$$7);
 		}
 		ZEPHIR_MM_RESTORE();
 		return Z_OBJ_P(this_ptr);