+++ tests/Extension/FcallTest.php
@@ -5,8 +5,8 @@
  *
  * (c) Phalcon Team <team@zephir-lang.com>
  *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
  */
 
 namespace Extension;
@@ -14,10 +14,6 @@ namespace Extension;
 use PHPUnit\Framework\TestCase;
 use Stub\Fcall;
 use Stub\Oo\PropertyAccess;
-
-class UserExample extends PropertyAccess
-{
-}
 
 class FcallTest extends TestCase
 {
@@ -120,13 +116,5 @@ class FcallTest extends TestCase
 
         $this->assertInstanceOf(\stdClass::class, \Stub\zephir_namespaced_method_with_type_casting(new \stdClass()));
         $this->assertInstanceOf(\stdClass::class, zephir_global_method_with_type_casting(new \stdClass()));
-    }
-
-    public function testIssue1851()
-    {
-        $example = new UserExample();
-        $example->setPrivateVariable('test');
-
-        $this->assertSame('test', $example->getPrivateVariable());
     }
 }